requirements.txt
Flask
flask-sqlalchemy
python-dotenv
requests
fuzzywuzzy
pandas
reportlab